Vending Machine Command Line Capstone/java/src/test/java/com/techelevator/view/InventoryTest.java
package com.techelevator.view;

import java.math.BigDecimal;
import java.math.RoundingMode;
import java.util.Map;
import java.util.TreeMap;

import org.junit.Before;
import org.junit.Test;
import org.junit.Assert;

public class InventoryTest {

	Inventory testInventory;
	Inventory testInventoryTwo; 

	@Before
	public void setup() {
		testInventory = new Inventory("test");
	}

	@Test
	public void get_item_name_functions_correctly() {
		String returnValue = testInventory.getItemName("A1");

		Assert.assertEquals("Candy Corn", returnValue);
	}

	@Test
	public void get_item_price_functions_correctly() {
		BigDecimal returnValue = testInventory.getItemPrice("B1");
		BigDecimal mountainBooPrice = new BigDecimal(3.44);
		Assert.assertEquals(mountainBooPrice.setScale(2, RoundingMode.HALF_DOWN), returnValue);
	}

	@Test
	public void correct_message_for_chips_displayed() {
		String returnValue = testInventory.getMessage("B2");

		Assert.assertEquals("Crunch Crunch, Yum!", returnValue);
	}

	@Test
	public void correct_message_for_candy_displayed() {
		String returnValue = testInventory.getMessage("A2");

		Assert.assertEquals("Munch Munch, Yum!", returnValue);
	}

	@Test
	public void correct_message_for_drink_displayed() {
		String returnValue = testInventory.getMessage("B1");

		Assert.assertEquals("Glug Glug, Yum!", returnValue);
	}

	@Test
	public void correct_message_for_gum_displayed() {
		String returnValue = testInventory.getMessage("C1");

		Assert.assertEquals("Chew Chew, Yum!", returnValue);
	}

	@Test
	public void decrement_item_count_by_one() {
		testInventory.decrementItemCount("C1");

		int expectedValue = 4;
		int returnValue = testInventory.codeCountMap.get("C1");

		Assert.assertEquals(expectedValue, returnValue);
	}

	// TEST always gives me 5.90 instead of 7.95!
	
//	@Test
//	public void vend_correct_item_update_number_of_items_sold_and_update_balance() {
//		Balance balanceForVend = new Balance("test");
//
//		testInventory.vend("A2", balanceForVend);
//
//		BigDecimal reesesPrice = testInventory.codeItemMap.get("A2").getPrice();
//		
//		
//		balanceForVend.subtractBalance(reesesPrice);
//		
//		BigDecimal returnValue = balanceForVend.getBalance();
//
//		int didItemSell = testInventory.codeItemMap.get("A2").getNumberSold();
//
//		Assert.assertEquals("Should add 1 to number sold", 1, didItemSell);
//		Assert.assertEquals("Vending item should subtract the price from your balance", 7.95, returnValue);
//	}

	@Test
	public void add_to_number_sold_works() {
		testInventory.addToNumberSold("A1");

		int expectedValue = 1;
		int returnValue = testInventory.codeItemMap.get("A1").getNumberSold();

		Assert.assertEquals(expectedValue, returnValue);
	}

}
